chore(ui): Rewrite IntegrationHealthService in TypeScript
